+++ src/hash/lohat.c
@@ -86,7 +86,7 @@ lohat_init_size(lohat_t *self, char size)
     lohat_store_t *store;
     uint64_t       len;
 
-    if (size > (sizeof(intptr_t) * 8)) {
+    if (((size_t)size) > (sizeof(intptr_t) * 8)) {
         abort();
     }
